feat: upd ResetForm
